Refactor Responder commands
Improve documentation just a bit
